Moved push() & pop() to new TurnStateContextCollection
